chore: Refactor and cleanup code formatting and comments
