+++ android/app/src/main/AndroidManifest.xml
@@ -37,15 +37,16 @@
             android:name=".activities.MainActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustPan">
+        </activity>
+        <activity
+            android:name=".activities.LoginActivity"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".activities.LoginActivity"
-            android:exported="true" />
     </application>
 
 </manifest>
