--- android/app/src/main/java/com/example/foodapp/activities/admin/AdminMainActivity.java
@@ -1,4 +1,0 @@
-package com.example.foodapp.activities.admin;
-
-public class AdminMainActivity {
-}
