--- android/app/src/main/java/com/example/foodapp/adapters/admin/AdminOrderAdapter.java
@@ -1,4 +1,0 @@
-package com.example.foodapp.adapters.admin;
-
-public class AdminOrderAdapter {
-}
